automation/ManejoWordExcel/requirements.txt
openpyxl==3.1.5
pandas==2.3.3
python-docx==1.2.0
